First draft for exercise number 1
Warning: This implementation is a draft. It will be cleared in the next
commit.

The goal is to read the jobs and professionals datasets to generate
category per continent. This implementation consists of reading
continent coordinates and store category names per professional_id.
This way, we have a fast way to obtain category name with a low memory
footprint as we are not storing the entire data.

After this preparation step, we are reading the job dataset.
As we have a lot of data in it AND as we have to do some
calculations (find if coordinate is in polygone), we are using Flow to
concurrently calculate and categorize all jobs coordinates.
The flat_map is used to find category and continent for each job. It
will also filters invalid data.
The most important part is the partition + reduce.
Partition will split flat_map result per continent name, we will have at
most 6 jobs running concurrently. Then, reduce will store each category
results per continent.

The last step is to print the table. It is done by using TableRex.