Add tests, clean categorize implementation
Creates a new module named `Ats.Dataset` to manipulate data of the 3
kinds of dataset:
- jobs
- professions
- continents

All functions are documented with small examples (that are doctests).

One extra test is created in test/ats/dataset_test.exs to test
calculation of number of categories per continent.

Test dataset are stored in `test/data` and are classified by `valid` or
`invalid` data.