Defines contract_type as an Enum
Defines values for contract type in GraphQL. There is no visible change
for the client on the returned value, but now it is typed and there is
no more surprise.

Creates an EctoEnum and enforces types constraint on Job's schema. Data
will be validated when Changeset.cast is called with `:contract_type`.

SQL migration converts all `contract_type` to lowercase to be supported
by Absinthe enum and EctoEnum.