Adds subscription.jobAdded
This is an experiment with subscriptions to see how it works.
It adds a subscription to mutation.createJob and expose it's results.
It can be useful for a real ATS to subscribe to new job to see them in
live.

The only things that's missing is how to test it.